+++ src/functions/background.js
@@ -305,6 +305,7 @@ chrome.runtime.onMessage.addListener(
                 });
                 return true;
             } break;
+
         }